add utils to get models
