+++ plot/subplot_visualizer.py
@@ -2,11 +2,11 @@ import sys
 import os
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
-from plot_average_energy import plot_average_energy
-from plot_environment_health import plot_environment_health
-from plot_agent_behavior import plot_agent_behavior
-from plot_energy_vs_health import plot_energy_vs_health
-from plot_tokens import plot_tokens
+from plot.plot_average_energy import plot_average_energy
+from plot.plot_environment_health import plot_environment_health
+from plot.plot_agent_behavior import plot_agent_behavior
+from plot.plot_energy_vs_health import plot_energy_vs_health
+from plot.plot_tokens import plot_tokens
 
 def subplot_visualizer(model):
     import matplotlib.pyplot as plt
